// app/_globalRedux/services/types/alerts.d.ts
export interface GetAlertsPopupsResponse {
	alert_id: string;
	alert_type: string;
	alertcount: string;
	datetime: string;
	msg: string;
	sys_service_id: string;
	vehicleno: string;
	status: string;
}